changed file structure in scss folder. Created new modules, copied old code. Vendor and Customer are now separate
