--- src/styles/css/vendor_profile.module.css
@@ -1,172 +1,0 @@
-.vendor_profile_container {
-  width: 100%;
-}
-
-.vendor_profile_container .vendor_header_container {
-  /* border: 3px solid purple; */
-  position: relative;
-  background-color: #00b2ed;
-  top: 0;
-  width: 100%;
-  height: 85px;
-  margin: 0 auto;
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-pack: justify;
-      -ms-flex-pack: justify;
-          justify-content: space-between;
-}
-
-.vendor_profile_container .vendor_header_container #hamburger_dropdown {
-  position: relative;
-  margin: 20px;
-  display: none;
-  -webkit-box-align: end;
-      -ms-flex-align: end;
-          align-items: flex-end;
-  height: -webkit-fit-content;
-  height: -moz-fit-content;
-  height: fit-content;
-  -webkit-transition: display 1s ease-in-out;
-  transition: display 1s ease-in-out;
-}
-
-@media only screen and (max-width: 1000px) {
-  .vendor_profile_container .vendor_header_container #hamburger_dropdown {
-    display: -webkit-box;
-    display: -ms-flexbox;
-    display: flex;
-    -webkit-box-orient: vertical;
-    -webkit-box-direction: normal;
-        -ms-flex-direction: column;
-            flex-direction: column;
-  }
-}
-
-.vendor_profile_container .vendor_header_container #hamburger_dropdown #closebtn {
-  position: relative;
-  right: 0;
-  cursor: pointer;
-}
-
-.vendor_profile_container .vendor_header_container #hamburger_dropdown #closebtn .line1,
-.vendor_profile_container .vendor_header_container #hamburger_dropdown #closebtn .line2,
-.vendor_profile_container .vendor_header_container #hamburger_dropdown #closebtn .line3 {
-  margin-top: 5px;
-  background-color: #fff;
-  width: 25px;
-  height: 3px;
-  display: block;
-  position: relative;
-  opacity: 1;
-  border-radius: 15%;
-  -webkit-transition: all 0.3s;
-  transition: all 0.3s;
-}
-
-.vendor_profile_container .vendor_header_container #hamburger_dropdown .hamburger_dropdown_links {
-  display: block;
-  background-color: #00b2ed;
-  z-index: 2;
-  margin-top: 5px;
-  padding: 10px;
-  height: -webkit-fit-content;
-  height: -moz-fit-content;
-  height: fit-content;
-}
-
-.vendor_profile_container .vendor_header_container #hamburger_dropdown .hamburger_dropdown_links p {
-  cursor: pointer;
-  position: relative;
-  top: 9px;
-  font-family: Raleway;
-  font-style: normal;
-  font-weight: 500;
-  font-size: 18px;
-  line-height: 170.5%;
-  -webkit-box-align: baseline;
-      -ms-flex-align: baseline;
-          align-items: baseline;
-  color: #fef2e4;
-}
-
-.vendor_profile_container .vendor_header_container #hamburger_dropdown .hamburger_dropdown_links p:hover {
-  color: lightblue;
-}
-
-.vendor_profile_container .vendor_header_container #hamburger_dropdown .no_drop {
-  display: none;
-}
-
-.vendor_profile_container .vendor_header_container h1 {
-  cursor: pointer;
-  width: -webkit-fit-content;
-  width: -moz-fit-content;
-  width: fit-content;
-  position: relative;
-  left: 75px;
-  -ms-flex-item-align: center;
-      -ms-grid-row-align: center;
-      align-self: center;
-  font-family: Raleway;
-  font-weight: 600;
-  font-size: 36px;
-  line-height: 42px;
-  color: #fef2e4;
-}
-
-.vendor_profile_container .vendor_header_container .header_links {
-  height: 100%;
-  width: 50%;
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -ms-flex-pack: distribute;
-      justify-content: space-around;
-  -webkit-box-align: center;
-      -ms-flex-align: center;
-          align-items: center;
-}
-
-@media only screen and (max-width: 1000px) {
-  .vendor_profile_container .vendor_header_container .header_links {
-    display: none;
-  }
-}
-
-.vendor_profile_container .vendor_header_container .header_links p {
-  cursor: pointer;
-  position: relative;
-  top: 9px;
-  font-family: Raleway;
-  font-style: normal;
-  font-weight: 500;
-  font-size: 18px;
-  line-height: 170.5%;
-  -webkit-box-align: baseline;
-      -ms-flex-align: baseline;
-          align-items: baseline;
-  color: #fef2e4;
-}
-
-.vendor_profile_container .vendor_header_container .header_links .header_food {
-  left: 62px;
-}
-
-@media only screen and (max-width: 1600px) {
-  .vendor_profile_container .vendor_header_container .header_links .header_food {
-    left: 0;
-  }
-}
-
-.vendor_profile_container .vendor_header_container .header_links .header_about {
-  left: 172px;
-}
-
-@media only screen and (max-width: 1600px) {
-  .vendor_profile_container .vendor_header_container .header_links .header_about {
-    left: 0;
-  }
-}
-/*# sourceMappingURL=vendor_profile.module.css.map */
